# docs/instant-messaging-api-spec.yml
openapi: 3.0.3
info:
  title: Instant Messaging API
  description: API for managing messaging channels
  version: 0.1.0
  # Antonio Paulino - 50512
  # Bernardo Pereira - 50493
servers:
  - url: http://localhost:8080/api
    description: Local server
...
paths:
...
  /api/auth/invitations:
    post:
      tags:
        - Authentication
      summary: Create an invitation
      security:
        - bearerAuth: []
      requestBody:
        description: Invitation details
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImInvitationCreationInputModel'
      responses:
        '201':
          description: Invitation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImInvitationOutputModel'
        '400':
            $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/channels/{channelID}/invitations:
    get:
      tags:
        - Invitations
      summary: Get a list of all channel invitations for a channel
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelID'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/GetCount'
        - $ref: '#/components/parameters/Sort'
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [ id, expiresAt, role ]
            default: expiresAt
        - name: after
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
            description: Get invitations after this ID
      responses:
        '200':
          description: List of channel invitations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelInvitationsOutputModel'
        '401':
            $ref: '#/components/responses/Unauthorized'
        '403':
            $ref: '#/components/responses/Forbidden'
        '404':
            $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'


    post:
      tags:
        - Invitations
      summary: Create a channel invitation
      security:
          - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelID'
      requestBody:
        description: Details to create a channel invitation
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelInvitationCreationInputModel'
      responses:
        '201':
          description: Invitation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelInvitationCreationOutputModel'
        '400':
            $ref: '#/components/responses/BadRequest'
        '401':
            $ref: '#/components/responses/Unauthorized'
        '403':
            $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/channels/{channelID}/invitations/{invitationID}:
    patch:
      tags:
        - Invitations
      summary: Update invite permissions and expiration date
      parameters:
        - $ref: '#/components/parameters/ChannelID'
        - $ref: '#/components/parameters/InvitationID'
      requestBody:
        description: Invitation status
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelInvitationUpdateInputModel'
      responses:
        '204':
          description: Invitation status updated successfully
        '400':
            $ref: '#/components/responses/BadRequest'
        '401':
            $ref: '#/components/responses/Unauthorized'
        '403':
            $ref: '#/components/responses/Forbidden'
        '404':
            $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Invitations
      summary: Delete a channel invitation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelID'
        - $ref: '#/components/parameters/InvitationID'
      responses:
        '204':
          description: Invitation deleted successfully
        '401':
            $ref: '#/components/responses/Unauthorized'
        '403':
            $ref: '#/components/responses/Forbidden'
        '404':
            $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/users/{userID}/invitations:
    get:
      tags:
        - Invitations
      summary: Get a list of all channel invitations for a user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserID'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/GetCount'
        - $ref: '#/components/parameters/Sort'
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [ id, expiresAt ]
            default: expiresAt
      responses:
        '200':
          description: List of channel invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChannelInvitationsOutputModel'
        '401':
            $ref: '#/components/responses/Unauthorized'
        '403':
            $ref: '#/components/responses/Forbidden'
        '404':
            $ref: '#/components/responses/NotFound'
        '429':
            $ref: '#/components/responses/TooManyRequests'
  /api/users/{userID}/invitations/{invitationID}:
    patch:
      tags:
        - Invitations
      summary: Accept or decline a channel invitation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserID'
        - $ref: '#/components/parameters/InvitationID'
      requestBody:
        description: Invitation status
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationAcceptInputModel'
      responses:
        '204':
          description: Invitation status updated successfully
        '400':
            $ref: '#/components/responses/BadRequest'
        '401':
            $ref: '#/components/responses/Unauthorized'
        '403':
            $ref: '#/components/responses/Forbidden'
        '404':
            $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    get:
        tags:
            - Invitations
        summary: Get a channel invitation's details
        parameters:
            - $ref: '#/components/parameters/UserID'
            - $ref: '#/components/parameters/InvitationID'
        responses:
            '200':
              description: Invitation retrieved successfully
              content:
                application/json:
                  schema:
                      $ref: '#/components/schemas/InvitationOutputModel'
            '401':
                $ref: '#/components/responses/Unauthorized'
            '403':
                $ref: '#/components/responses/Forbidden'
            '404':
                $ref: '#/components/responses/NotFound'
            '429':
              $ref: '#/components/responses/TooManyRequests'
components:
  schemas:
...
    ImInvitationCreationInputModel:
      type: object
      properties:
        expiresAt:
          type: string
          format: date-time
    ImInvitationOutputModel:
      type: object
      properties:
        token:
          type: string
          format: uuid
...
    ChannelInvitationCreationInputModel:
      type: object
      properties:
        invitee:
          type: integer
        role:
          type: string
          enum: [ owner, member, guest ]
        expiresAt:
          type: string
          format: date-time
      required:
        - expiresAt
        - role
    ChannelInvitationUpdateInputModel:
      type: object
      properties:
        role:
          type: string
          enum: [ member, guest ]
        expiresAt:
          type: string
          format: date-time
    ChannelInvitationsOutputModel:
      type: object
      properties:
        invitations:
          type: array
          items:
            $ref: '#/components/schemas/InvitationOutputModel'
    InvitationOutputModel:
      type: object
      properties:
        invitationID:
          type: integer
        role:
          type: string
          enum: [ member, guest ]
        expirationDate:
          type: string
          format: date-time
    InvitationAcceptInputModel:
      type: object
      properties:
        status:
          type: string
          enum: [ accepted, rejected ]
      required:
        - status
    Problem:
      description: Error response
      type: object
      properties:
        type:
          type: string
        title:
          type: string
          format: uri
        status:
          type: integer
        detail:
          type: string
    InputValidationProblem:
      description: Validation error for Body, Query, or Path parameters
      type: object
      properties:
        type:
          type: string
        title:
          type: string
          format: uri
        status:
          type: integer
        detail:
          type: string
        errors:
          type: array
          items:
            type: string
...
    ChannelInvitationCreationOutputModel:
      type: object
      properties:
        id:
          type: integer
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/Problem'
              - $ref: '#/components/schemas/InputValidationProblem'
    Unauthorized:
        description: Unauthorized
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Problem'
    Forbidden:
        description: Forbidden
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/Problem'
    TooManyRequests:
        description: Too many requests
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/Problem'
    NotFound:
        description: Not found
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/Problem'
...
  parameters:
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 50
        maximum: 100
        minimum: 1
...
    GetCount:
      name: getCount
      in: query
      required: false
      schema:
        type: boolean
        default: true
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        enum: [ asc, desc ]
        default: asc
    UserID:
      name: userID
      in: path
      required: true
      schema:
        type: number
    ChannelID:
      name: channelID
      in: path
      required: true
      schema:
        type: number
...
    InvitationID:
      name: invitationID
      in: path
      required: true
      schema:
        type: number
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID